requirements.txt
amqp==5.0.9
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
billiard==3.6.4.0
black==22.3.0
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.2
django-extensions==3.1.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
executing==0.8.3
filelock==3.5.1
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
ipdb==0.13.9
ipython==8.1.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
schedule==1.1.0
sentry-sdk==1.5.12
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
split-settings==1.0.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.13.1
wcwidth==0.2.5
wrapt==1.13.3
